+++ Po.py
@@ -29,4 +29,4 @@ class Po( gi.Game ):
 
 	iPosB1 = ( int( width * ( 3 / 8 )), int( gi.Game.size_b ))
 
-	iPosS1 = ( width * ( 1 / 2 ), height * ( 1 / 2 ))
+	iPosS1 = ( int( width * ( 1 / 2 )), int( height * ( 1 / 2 )))
